+++ processor/src/trace/mod.rs
@@ -17,7 +17,10 @@ use winter_prover::{crypto::RandomCoin, EvaluationFrame, Trace, TraceLayout};
 use vm_core::StarkField;
 
 mod utils;
-pub use utils::{build_lookup_table_row_values, AuxColumnBuilder, LookupTableRow, TraceFragment};
+pub use utils::{
+    build_lookup_table_row_values, AuxColumnBuilder, ChipletsLengths, LookupTableRow,
+    TraceFragment, TraceLenSummary,
+};
 
 mod decoder;
 
@@ -54,6 +57,7 @@ pub struct ExecutionTrace {
     aux_trace_hints: AuxTraceHints,
     program_info: ProgramInfo,
     stack_outputs: StackOutputs,
+    trace_len_summary: TraceLenSummary,
 }
 
 impl ExecutionTrace {
@@ -80,7 +84,7 @@ impl ExecutionTrace {
         // create a new program info instance with the underlying kernel
         let kernel = process.kernel().clone();
         let program_info = ProgramInfo::new(program_hash, kernel);
-        let (main_trace, aux_trace_hints) = finalize_trace(process, rng);
+        let (main_trace, aux_trace_hints, trace_len_summary) = finalize_trace(process, rng);
 
         Self {
             meta: Vec::new(),
@@ -89,6 +93,7 @@ impl ExecutionTrace {
             aux_trace_hints,
             program_info,
             stack_outputs,
+            trace_len_summary,
         }
     }
 
@@ -143,6 +148,11 @@ impl ExecutionTrace {
         self.main_trace.num_rows()
     }
 
+    /// Returns a summary of the lengths of main, range and chiplet traces.
+    pub fn trace_len_summary(&self) -> &TraceLenSummary {
+        &self.trace_len_summary
+    }
+
     // HELPER METHODS
     // --------------------------------------------------------------------------------------------
 
@@ -164,7 +174,9 @@ impl ExecutionTrace {
     }
 
     #[cfg(test)]
-    pub fn test_finalize_trace<A>(process: Process<A>) -> (Vec<Vec<Felt>>, AuxTraceHints)
+    pub fn test_finalize_trace<A>(
+        process: Process<A>,
+    ) -> (Vec<Vec<Felt>>, AuxTraceHints, TraceLenSummary)
     where
         A: AdviceProvider,
     {
@@ -263,7 +275,10 @@ impl Trace for ExecutionTrace {
 /// - Inserting random values in the last row of all columns. This helps ensure that there
 ///   are no repeating patterns in each column and each column contains a least two distinct
 ///   values. This, in turn, ensures that polynomial degrees of all columns are stable.
-fn finalize_trace<A>(process: Process<A>, mut rng: RpoRandomCoin) -> (Vec<Vec<Felt>>, AuxTraceHints)
+fn finalize_trace<A>(
+    process: Process<A>,
+    mut rng: RpoRandomCoin,
+) -> (Vec<Vec<Felt>>, AuxTraceHints, TraceLenSummary)
 where
     A: AdviceProvider,
 {
@@ -292,6 +307,10 @@ where
         trace_len >= MIN_TRACE_LEN,
         "trace length must be at least {MIN_TRACE_LEN}, but was {trace_len}",
     );
+
+    // get the lengths of the traces: main, range, and chiplets
+    let trace_len_summary =
+        TraceLenSummary::new(clk as usize, range_table_len, ChipletsLengths::new(&chiplets));
 
     // combine all trace segments into the main trace
     let system_trace = system.into_trace(trace_len, NUM_RAND_ROWS);
@@ -324,5 +343,5 @@ where
         chiplets: chiplets_trace.aux_builder,
     };
 
-    (trace, aux_trace_hints)
+    (trace, aux_trace_hints, trace_len_summary)
 }