+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [BREAKING] Added options to specify maximum number of cycles and expected number of cycles for a program (#998).
 - Improved handling of invalid/incomplete parameters in `StackOutputs` constructors (#1010).
 - Allowed the assembler to produce programs with "phantom" calls (#1019).
+- Added `TraceLenSummary` struct which holds information about traces lengths to the `ExecutionTrace` (#1029).
 
 ## 0.6.1 (2023-06-29)